Обновление

+++ code/datums/beam.dm
@@ -7,7 +7,6 @@
 	var/icon
 	var/icon_state = "" //icon state of the main segments of the beam
 	var/max_distance = 0
-	var/endtime = 0
 	var/sleep_time = 3
 	var/finished = 0
 	var/target_oldloc = null
@@ -15,14 +14,17 @@
 	var/static_beam = 0
 	var/beam_type = /obj/effect/ebeam //must be subtype
 	var/layer = null
+	var/timing_id = null
+	var/recalculating = FALSE
 
+	var/obj/effect/ebeam/visuals //what we add to the ebeam's visual contents. never gets deleted on redrawing.
 
-/datum/beam/New(beam_origin,beam_target,beam_icon='icons/effects/beam.dmi',beam_icon_state="b_beam",time=50,maxdistance=10,btype = /obj/effect/ebeam,beam_layer)
-	endtime = world.time+time
+/datum/beam/New(beam_origin,beam_target,beam_icon='icons/effects/beam.dmi',beam_icon_state="b_beam",time=50,maxdistance=10,btype = /obj/effect/ebeam, beam_sleep_time=3, beam_layer)
 	origin = beam_origin
 	origin_oldloc =	get_turf(origin)
 	target = beam_target
 	target_oldloc = get_turf(target)
+	sleep_time = beam_sleep_time
 	if(origin_oldloc == origin && target_oldloc == target)
 		static_beam = 1
 	max_distance = maxdistance
@@ -31,11 +33,23 @@
 	icon_state = beam_icon_state
 	beam_type = btype
 	layer = beam_layer
-
+	if(time < INFINITY)
+		addtimer(CALLBACK(src,.proc/End), time)
 
 /datum/beam/proc/Start()
+	visuals = new beam_type()
+	visuals.icon = icon
+	visuals.icon_state = icon_state
 	Draw()
-	while(!finished && origin && target && world.time < endtime && get_dist(origin,target)<max_distance && origin.z == target.z)
+	recalculate_in(sleep_time)
+
+/datum/beam/proc/recalculate()
+	if(recalculating)
+		recalculate_in(sleep_time)
+		return
+	recalculating = TRUE
+	timing_id = null
+	if(origin && target && get_dist(origin,target)<max_distance && origin.z == target.z)
 		var/origin_turf = get_turf(origin)
 		var/target_turf = get_turf(target)
 		if(!static_beam && (origin_turf != origin_oldloc || target_turf != target_oldloc))
@@ -43,30 +57,46 @@
 			target_oldloc = target_turf
 			Reset()
 			Draw()
-		sleep(sleep_time)
+		after_calculate()
+		recalculating = FALSE
+	else
+		End()
 
-	qdel(src)
+/datum/beam/proc/afterDraw()
+	return
 
+/datum/beam/proc/recalculate_in(time)
+	if(timing_id)
+		deltimer(timing_id)
+	timing_id = addtimer(CALLBACK(src, .proc/recalculate), time, TIMER_STOPPABLE)
 
-/datum/beam/proc/End()
-	finished = 1
+/datum/beam/proc/after_calculate()
+	if((sleep_time == null) || finished)	//Does not automatically recalculate.
+		return
+	if(isnull(timing_id))
+		timing_id = addtimer(CALLBACK(src, .proc/recalculate), sleep_time, TIMER_STOPPABLE)
 
+/datum/beam/proc/End(destroy_self = TRUE)
+	finished = TRUE
+	if(!isnull(timing_id))
+		deltimer(timing_id)
+	if(!QDELETED(src) && destroy_self)
+		qdel(src)
 
 /datum/beam/proc/Reset()
 	for(var/obj/effect/ebeam/B in elements)
 		qdel(B)
-
+	elements.Cut()
 
 /datum/beam/Destroy()
 	Reset()
+	qdel(visuals)
 	target = null
 	origin = null
 	return ..()
 
-
 /datum/beam/proc/Draw()
 	var/Angle = round(Get_Angle(origin,target))
-
 	var/matrix/rot_matrix = matrix()
 	rot_matrix.Turn(Angle)
 
@@ -77,20 +107,22 @@
 	var/length = round(sqrt((DX)**2+(DY)**2)) //hypotenuse of the triangle formed by target and origin's displacement
 
 	for(N in 0 to length-1 step 32)//-1 as we want < not <=, but we want the speed of X in Y to Z and step X
+		if(QDELETED(src) || finished)
+			break
 		var/obj/effect/ebeam/X = new beam_type(origin_oldloc)
 		X.owner = src
 		if(layer)
 			X.layer = layer
-		elements |= X
+		elements += X
 
-		//Assign icon, for main segments it's base_icon, for the end, it's icon+icon_state
-		//cropped by a transparent box of length-N pixel size
-		if(N+32>length)
-			var/icon/II = new(icon, icon_state)
-			II.DrawBox(null,1,(length-N),32,32)
+		//Assign our single visual ebeam to each ebeam's vis_contents
+		//ends are cropped by a transparent box icon of length-N pixel size laid over the visuals obj
+		if(N+32>length) //went past the target, needs to be cut short
+			var/icon/II = new(icon, icon_state) //the way to keep this the same as the vis_contents is unreasonable right now, maybe in the far future.
+			II.DrawBox(null,1,(length-N),32,32)//anyway we cut the icon on the ebeam to end at the target instead of overshooting
 			X.icon = II
 		else
-			X.icon = base_icon
+			X.vis_contents += visuals
 		X.transform = rot_matrix
 
 		//Calculate pixel offsets (If necessary)
@@ -108,31 +140,35 @@
 		//Position the effect so the beam is one continous line
 		var/a
 		if(abs(Pixel_x)>32)
-			a = Pixel_x > 0 ? round(Pixel_x / 32) : CEIL(Pixel_x / 32)
+			a = Pixel_x > 0 ? round(Pixel_x/32) : CEILING(Pixel_x/32, 1)
 			X.x += a
 			Pixel_x %= 32
 		if(abs(Pixel_y)>32)
-			a = Pixel_y > 0 ? round(Pixel_y / 32) : CEIL(Pixel_y / 32)
+			a = Pixel_y > 0 ? round(Pixel_y/32) : CEILING(Pixel_y/32, 1)
 			X.y += a
 			Pixel_y %= 32
 
 		X.pixel_x = Pixel_x
 		X.pixel_y = Pixel_y
 		CHECK_TICK
-
+	afterDraw()
 
 /obj/effect/ebeam
-	mouse_opacity = 0
-	anchored = 1
+	mouse_opacity = MOUSE_OPACITY_TRANSPARENT
+	anchored = TRUE
 	var/datum/beam/owner
-
 
 /obj/effect/ebeam/Destroy()
 	owner = null
 	return ..()
 
+/obj/effect/ebeam/singularity_pull()
+	return
 
-/atom/proc/Beam(atom/BeamTarget,icon_state="b_beam",icon='icons/effects/beam.dmi',time=50, maxdistance=10,beam_type=/obj/effect/ebeam,beam_layer=null)
-	var/datum/beam/newbeam = new(src,BeamTarget,icon,icon_state,time,maxdistance,beam_type,beam_layer)
-	INVOKE_ASYNC(newbeam, /datum/beam.proc/Start)
+/obj/effect/ebeam/singularity_act()
+	return
+
+/atom/proc/Beam(atom/BeamTarget,icon_state="b_beam",icon='icons/effects/beam.dmi',time=50, maxdistance=10,beam_type=/obj/effect/ebeam,beam_sleep_time = 3,beam_layer)
+	var/datum/beam/newbeam = new(src,BeamTarget,icon,icon_state,time,maxdistance,beam_type,beam_sleep_time,beam_layer)
+	INVOKE_ASYNC(newbeam, /datum/beam/.proc/Start)
 	return newbeam
